clean up immutable js from most of components and logic
