+++ app/modules/editor/components/StaticForm/StaticForm.styled.tsx
