+++ app/modules/editor/scenes/StaticEditorScene.tsx
@@ -1,6 +1,6 @@
 import React, { FC } from 'react';
-import SceneWrapper from '../../components/SceneWrapper/SceneWrapper';
-import StaticForm from './components/StaticForm/StaticForm';
+import SceneWrapper from '../../../components/SceneWrapper/SceneWrapper';
+import StaticForm from '../components/StaticForm/StaticForm';
 
 const StaticEditorScene: FC = () => (
   <SceneWrapper>
